+++ R/hooks-md.R
@@ -69,9 +69,8 @@ render_jekyll = function(highlight = c('pygments', 'prettify', 'none'), extra = 
   if (hi == 'none') return()
   switch(hi, pygments = {
     hook.r = function(x, options) {
-      x = paste(c(hilight_source(x, 'markdown', options), ''), collapse = '\n')
       str_c('\n\n{% highlight ', tolower(options$engine), if (extra != '') ' ', extra, ' %}\n',
-            x, '{% endhighlight %}\n\n')
+            x, '\n{% endhighlight %}\n\n')
     }
     hook.t = function(x, options) str_c('\n\n{% highlight text %}\n', x, '{% endhighlight %}\n\n')
   }, prettify = {
@@ -82,6 +81,8 @@ render_jekyll = function(highlight = c('pygments', 'prettify', 'none'), extra = 
     hook.t = function(x, options) str_c('\n\n<pre><code>', escape_html(x), '</code></pre>\n\n')
   })
   hook.o = function(x, options) if (output_asis(x, options)) x else hook.t(x, options)
-  knit_hooks$set(source = hook.r, output = hook.o, warning = hook.t,
-                 error = hook.t, message = hook.t)
+  knit_hooks$set(source = function(x, options) {
+    x = paste(hilight_source(x, 'markdown', options), collapse = '\n')
+    hook.r(x, options)
+  }, output = hook.o, warning = hook.t, error = hook.t, message = hook.t)
 }