+++ NEWS.md
@@ -24,6 +24,8 @@
 
 - fixed #604: `kable()` did not work on data frames/matrices of one row (thanks, Kevin Ushey)
 
+- fixes the bug reported at http://stackoverflow.com/q/18992260/559676; `render_jekyll('prettify')` should have pasted the source code lines into one character string
+
 ## MAJOR CHANGES
 
 - when the chunk option `cache=2`, the recorded plots (i.e. display lists) will no longer be cached, and the figure files will be assumed to exist the next time the document is compiled, otherwise the cache will be purged and the chunk needs to be recomputed; this will save some disk space and avoid bugs like #588 (thanks, @knokknok)